requirements.txt
fastapi[standard]==0.115.6
pymongo[srv]==4.10.1 # Mongo db driver
oauth2client==4.1.3 
google-cloud-texttospeech==2.21.1       # Obtener el audio con textos
google-generativeai==0.8.3      # Gemini. 
python-dotenv==1.0.1
fal-client==0.5.6   # to generate image with this client
pillow==10.3.0      # to help image manipulation
pydantic-ai==0.0.19
dataclouder-conversation-ai-cards==0.0.4
dataclouder-mongo==0.0.3  # MongoDB utilities
